always trying to rm tmp/sra/.sra - replace with if statement

+++ rules/download.smk
@@ -33,15 +33,12 @@ rule fastqdump:
 		prefetch --max-size 1024000000 {params.accession}
 		fastq-dump --split-files --gzip --defline-seq '@$ac-$sn/$ri' --defline-qual '+' {params.accession}  
 
-#		# cleanup
-#		rm tmp/sra/{params.accession}.sra
-
 		#rename to expected output files
 		mv {params.accession}_1.fastq.gz {output.f}
 		mv {params.accession}_2.fastq.gz {output.r}
 
 		# cleanup
-		rm tmp/sra/{params.accession}.sra
+		if [[ -f tmp/sra/{params.accession}.sra ]]; then rm tmp/sra/{params.accession}.sra; fi
 		"""
 
 rule prep_local_reads:
